added namespace and changed to css modules
