+++ src/assets/css/authentication/MobileLoginContainer.module.css
@@ -1,5 +1,5 @@
-/* MobileLoginContainer.css */
-.mobile-login-container {
+/* MobileLoginContainer.module.css */
+.container {
     display: flex;
     justify-content: center;
     align-items: center;
@@ -14,7 +14,7 @@
 /* For small mobile phones */
 @media (max-width: 480px) {
     /* styles for screens smaller than 600px */
-    .login-form {
+    .loginForm {
         width: 100%; /* Full width on mobile */
         min-width: 320px; /* But not too small */
         padding: 20px;
@@ -27,7 +27,7 @@
 /* For large mobile phones */
 @media (min-width: 481px) and (max-width: 768px) {
     /* styles for screens smaller than 600px */
-    .login-form {
+    .loginForm {
         width: 100%; /* Full width on mobile */
         padding: 20px;
         background-color: white;
@@ -39,7 +39,7 @@
 /* For tablets */
 @media (min-width: 768px) and (max-width: 1024px) {
     /* styles for screens between 601px and 1024px */
-    .login-form {
+    .loginForm {
         width: 100%; /* Full width on mobile */
         padding: 20px;
         background-color: white;
