--- src/components/MobileLoginContainer.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import '../assets/css/MobileLoginContainer.css'; // Import mobile-specific CSS
-import reactLogo from '../assets/img/react.svg';
-
-const MobileLoginContainer = () => {
-  return (
-    <div className="mobile-login-container">
-      <div className="login-form">
-        {/* Your login form goes here */}
-      </div>
-    </div>
-  );
-}
-
-export default MobileLoginContainer;
