--- src/components/WebLoginContainer.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import '../assets/css/WebLoginContainer.css'; // Import web-specific CSS
-import reactLogo from '../assets/img/react.svg'
-
-const WebLoginContainer = () => {
-  return (
-    <div className="web-login-container">
-    <div className="login-form">
-      {/* Your login form goes here */}
-    </div>
-  </div>
-  );
-}
-
-export default WebLoginContainer;
