+++ src/components/authentication/MobileLoginContainer.tsx
@@ -1,0 +1,13 @@
+import styles from '../../assets/css/authentication/MobileLoginContainer.module.css'; // Import mobile-specific CSS
+
+const MobileLoginContainer = () => {
+  return (
+    <div className={styles.container}>
+      <div className={styles.loginForm}>
+        {/* Your login form goes here */}
+      </div>
+    </div>
+  );
+}
+
+export default MobileLoginContainer;
