+++ src/components/authentication/WebLoginContainer.tsx
@@ -1,0 +1,13 @@
+import styles from '../../assets/css/authentication/WebLoginContainer.module.css'; // Import web-specific CSS
+
+const WebLoginContainer = () => {
+  return (
+    <div className={styles.container}>
+      <div className={styles.loginForm}>
+        {/* Your login form goes here */}
+      </div>
+    </div>
+  );
+}
+
+export default WebLoginContainer;
